requirements.txt
asgiref==3.7.2
Django==5.0.3
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
colorama==0.4.6
cryptography==39.0.1
dnspython==2.3.0
emailfinder==0.3.0b0
flextls==0.3
future==0.18.3
google==3.0.0
idna==3.4
prettyprint==0.1.5
prompt-toolkit==3.0.36
pycparser==2.21
pyfiglet==0.8.post1
pyOpenSSL==23.0.0
pyperclip==1.8.2
python-whois==0.8.0
requests==2.28.2
shodan==1.28.0
six==1.16.0
soupsieve==2.3.2.post1
sslscan==0.5.1
urllib3==1.26.14
wcwidth==0.2.6
XlsxWriter==3.0.7